refactored network factory
